Updating Boolean type to be intptr_t so get/set funcitions work better.
jeff


git-svn-id: http://anonsvn.internet2.edu/svn/I2util/trunk@153 af504c07-53ac-4b46-a7f1-a0422916b9a2

+++ I2util/util.h
@@ -30,7 +30,7 @@
 #define	__attribute__(x)
 #endif
 
-typedef unsigned int    I2Boolean;
+typedef intptr_t    I2Boolean;
 
 #ifndef	False
 #define	False	(0)
